split window into tabs
